+++ lib/pylon.dart
@@ -1,10 +1,11 @@
 library pylon;
 
+export 'package:pylon_codec/pylon_codec.dart';
+
 export 'src/cluster.dart';
 export 'src/future.dart';
 export 'src/mutable_pylon.dart';
+export 'src/port.dart';
 export 'src/pylon.dart';
 export 'src/stream.dart';
-export 'src/port.dart';
-export 'src/codec.dart';
 export 'src/util/extensions.dart';